feat: 주입된 service 와 request 의 타입이 맞지 않는 경우 테스트 케이스 추가
